Learner generating suffixes for Gen and Loc of Polish words being in Nom
Two ways of getting input data: either put some long Polish texts into
polish_texts/ directory (I suggest a few books)
or dump whole Polish dictionary from Aspell. All the words are fed into
Morfologik that tries to find appropriate base form of every word.
If successful, then the common part of the two variants of the word
(in nominative form and inflected) is removed and the pair of suffixes
is stored in the map. The results of either of these algorithms
are supplemented by a list of manually inserted
'nominativus -> inflected' suffixes